backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
google-genai>=1.0.0
supabase>=2.0.0
httpx>=0.28.0,<0.29.0
pydantic[email]>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.25.0
ruff>=0.9.0
